Removed APEUnitWebViewDelegateV2
